自定义函数

+++ spark_sql/src/main/scala/com/mayishijie/makeDF/RddToDF.scala
@@ -44,4 +44,4 @@ object RddToDF {
 
 }
 
-case class Person(name:String,age:Int)
+case class Person(name:String,age:Long)
